adding some functionality and design

+++ app/helpers/application_helper.rb
@@ -6,6 +6,7 @@ module ApplicationHelper
       links += "<li> #{link_to('Edit Profile', edit_users_path )}</li>" +
       links += "<li> #{link_to('Logout ' + @authenticated.username, user_login_path, :method => 'delete', :confirm => 'Are you sure?')}</li>"
     elsif @company_authenticated.present?
+      links += "<li> #{link_to(@company_authenticated.name + ' Profile', company_path(@company_authenticated))}</li>" +
       links += "<li> #{link_to('+ New Product', new_product_path)}</li>" +
       links += "<li> #{link_to('Logout ' + @company_authenticated.name, company_login_path, :method => 'delete', :confirm => 'You are about to Logout. Do you wish to continue?')}</li>"
     else
@@ -19,7 +20,7 @@ module ApplicationHelper
 
   def edit_product_menu
     links = ""
-    if @company_authenticated.present?
+    if @company_authenticated.present? && @company_authenticated.id == @product.company_id
       #links to edit and delete product
       links += "#{link_to('Edit', edit_product_path(@product))}" +
       links += "#{button_to('Delete', @product, :method => 'delete')}" +
